chore: separate builder.io components
